requirements.txt
Flask==3.0.1
Flask_Cors==4.0.0
protobuf==3.20.3
python-dotenv==1.0.1
